Enable CompletionStage/CompletableFuture interop on native (#10358)
