--- core/native/src/main/scala/zio/ZIOPlatformSpecificJVM.scala
@@ -1,3 +1,0 @@
-package zio
-
-private[zio] trait ZIOPlatformSpecificJVM
